+++ code/model.py
@@ -1,0 +1,56 @@
+import transformers
+import pytorch_lightning as pl
+import torchmetrics
+import torch
+
+
+class Model(pl.LightningModule):
+    def __init__(self, model_name, lr):
+        super().__init__()
+        self.save_hyperparameters()
+
+        self.model_name = model_name
+        self.lr = lr
+
+        self.plm = transformers.AutoModelForSequenceClassification.from_pretrained(
+            pretrained_model_name_or_path=model_name, num_labels=1)
+        self.loss_func = torch.nn.L1Loss()
+
+    def forward(self, x, attention_mask):
+        x = self.plm(x, attention_mask=attention_mask)['logits']
+
+        return x
+
+    def training_step(self, batch, batch_idx):
+        x, attention_mask, y = batch
+        logits = self(x, attention_mask)
+        loss = self.loss_func(logits, y.float())
+        self.log("train_loss", loss)
+
+        return loss
+
+    def validation_step(self, batch, batch_idx):
+        x, attention_mask, y = batch
+        logits = self(x, attention_mask)
+        loss = self.loss_func(logits, y.float())
+        self.log("val_loss", loss)
+
+        self.log("val_pearson", torchmetrics.functional.pearson_corrcoef(logits.squeeze(), y.squeeze()))
+
+        return loss
+
+    def test_step(self, batch, batch_idx):
+        x, attention_mask, y = batch
+        logits = self(x, attention_mask)
+
+        self.log("test_pearson", torchmetrics.functional.pearson_corrcoef(logits.squeeze(), y.squeeze()))
+
+    def predict_step(self, batch, batch_idx):
+        x, attention_mask = batch
+        logits = self(x, attention_mask)
+
+        return logits.squeeze()
+
+    def configure_optimizers(self):
+        optimizer = torch.optim.AdamW(self.parameters(), lr=self.lr)
+        return optimizer
