+++ code/train.py
@@ -5,217 +5,24 @@ import pandas as pd
 
 from tqdm.auto import tqdm
 
-import transformers
+from transformers import TrainingArguments, Trainer
 import torch
-import torchmetrics
 import pytorch_lightning as pl
 
 import optuna
-import os
-os.environ["TOKENIZERS_PARALLELISM"] = "false"
+from dotenv import load_dotenv
 
-# seed 고정
+from .dataloader import Dataloader
+from .model import Model
+
 torch.manual_seed(0)
 torch.cuda.manual_seed(0)
 torch.cuda.manual_seed_all(0)
 random.seed(0)
 
-# Define the objective function for Optuna
-def objective(trial, args):
-    # Suggest hyperparameters to tune
-    batch_size = trial.suggest_categorical('batch_size', [8, 16])
-    learning_rate = trial.suggest_float('learning_rate', 1e-6, 1e-4, log=True)
-    max_epoch = trial.suggest_int('max_epoch', 1, 30)
-    
-    dataloader = Dataloader(
-        args.model_name, 
-        batch_size,
-        args.shuffle,
-        args.train_path,
-        args.dev_path,
-        args.test_path,
-        args.predict_path,
-        args.num_workers
-    )
-    
-    model = Model(args.model_name, lr=learning_rate)
-    
-    trainer = pl.Trainer(
-        accelerator="gpu", devices=1, 
-        max_epochs=max_epoch,
-        log_every_n_steps=1,
-        logger=False,
-        enable_checkpointing=False
-    )
-    
-    trainer.fit(model=model, datamodule=dataloader)
-    val_metrics = trainer.validate(model=model, datamodule=dataloader, verbose=False)
-    val_loss = val_metrics[0]['val_loss']
-    return val_loss
-
-
-class Dataset(torch.utils.data.Dataset):
-    def __init__(self, inputs, targets=[]):
-        self.inputs = inputs
-        self.targets = targets
-
-    # 학습 및 추론 과정에서 데이터를 1개씩 꺼내오는 곳
-    def __getitem__(self, idx):
-        # 정답이 있다면 else문을, 없다면 if문을 수행합니다
-        if len(self.targets) == 0:
-            return torch.tensor(self.inputs[idx])
-        else:
-            return torch.tensor(self.inputs[idx]), torch.tensor(self.targets[idx])
-
-    # 입력하는 개수만큼 데이터를 사용합니다
-    def __len__(self):
-        return len(self.inputs)
-
-
-class Dataloader(pl.LightningDataModule):
-    def __init__(self, model_name, batch_size, shuffle, train_path, dev_path, test_path, predict_path, num_workers):
-        super().__init__()
-        self.model_name = model_name
-        self.batch_size = batch_size
-        self.shuffle = shuffle
-
-        self.train_path = train_path
-        self.dev_path = dev_path
-        self.test_path = test_path
-        self.predict_path = predict_path
-        self.num_workers = num_workers
-
-        self.train_dataset = None
-        self.val_dataset = None
-        self.test_dataset = None
-        self.predict_dataset = None
-
-        self.tokenizer = transformers.AutoTokenizer.from_pretrained(model_name, max_length=160)
-        self.target_columns = ['label']
-        self.delete_columns = ['id']
-        self.text_columns = ['sentence_1', 'sentence_2']
-
-    def tokenizing(self, dataframe):
-        data = []
-        for idx, item in tqdm(dataframe.iterrows(), desc='tokenizing', total=len(dataframe)):
-            # 두 입력 문장을 [SEP] 토큰으로 이어붙여서 전처리합니다.
-            text = '[SEP]'.join([item[text_column] for text_column in self.text_columns])
-            outputs = self.tokenizer(text, add_special_tokens=True, padding='max_length', truncation=True)
-            data.append(outputs['input_ids'])
-        return data
-
-    def preprocessing(self, data):
-        # 안쓰는 컬럼을 삭제합니다.
-        data = data.drop(columns=self.delete_columns)
-
-        # 타겟 데이터가 없으면 빈 배열을 리턴합니다.
-        try:
-            targets = data[self.target_columns].values.tolist()
-        except:
-            targets = []
-        # 텍스트 데이터를 전처리합니다.
-        inputs = self.tokenizing(data)
-
-        return inputs, targets
-
-    def setup(self, stage='fit'):
-        if stage == 'fit':
-            # 학습 데이터와 검증 데이터셋을 호출합니다
-            train_data = pd.read_csv(self.train_path)
-            val_data = pd.read_csv(self.dev_path)
-
-            # 학습데이터 준비
-            train_inputs, train_targets = self.preprocessing(train_data)
-
-            # 검증데이터 준비
-            val_inputs, val_targets = self.preprocessing(val_data)
-
-            # train 데이터만 shuffle을 적용해줍니다, 필요하다면 val, test 데이터에도 shuffle을 적용할 수 있습니다
-            self.train_dataset = Dataset(train_inputs, train_targets)
-            self.val_dataset = Dataset(val_inputs, val_targets)
-        else:
-            # 평가데이터 준비
-            test_data = pd.read_csv(self.test_path)
-            test_inputs, test_targets = self.preprocessing(test_data)
-            self.test_dataset = Dataset(test_inputs, test_targets)
-
-            predict_data = pd.read_csv(self.predict_path)
-            predict_inputs, predict_targets = self.preprocessing(predict_data)
-            self.predict_dataset = Dataset(predict_inputs, [])
-
-    def train_dataloader(self):
-        return torch.utils.data.DataLoader(self.train_dataset, batch_size=self.batch_size, num_workers=self.num_workers, shuffle=self.shuffle)
-
-    def val_dataloader(self):
-        return torch.utils.data.DataLoader(self.val_dataset, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def test_dataloader(self):
-        return torch.utils.data.DataLoader(self.test_dataset, batch_size=self.batch_size, num_workers=self.num_workers)
-
-    def predict_dataloader(self):
-        return torch.utils.data.DataLoader(self.predict_dataset, batch_size=self.batch_size, num_workers=self.num_workers)
-
-
-class Model(pl.LightningModule):
-    def __init__(self, model_name, lr):
-        super().__init__()
-        self.save_hyperparameters()
-
-        self.model_name = model_name
-        self.lr = lr
-
-        # 사용할 모델을 호출합니다.
-        self.plm = transformers.AutoModelForSequenceClassification.from_pretrained(
-            pretrained_model_name_or_path=model_name, num_labels=1)
-        # Loss 계산을 위해 사용될 L1Loss를 호출합니다.
-        self.loss_func = torch.nn.L1Loss()
-
-    def forward(self, x):
-        x = self.plm(x)['logits']
-
-        return x
-
-    def training_step(self, batch, batch_idx):
-        x, y = batch
-        logits = self(x)
-        loss = self.loss_func(logits, y.float())
-        self.log("train_loss", loss)
-
-        return loss
-
-    def validation_step(self, batch, batch_idx):
-        x, y = batch
-        logits = self(x)
-        loss = self.loss_func(logits, y.float())
-        self.log("val_loss", loss)
-
-        self.log("val_pearson", torchmetrics.functional.pearson_corrcoef(logits.squeeze(), y.squeeze()))
-
-        return loss
-
-    def test_step(self, batch, batch_idx):
-        x, y = batch
-        logits = self(x)
-
-        self.log("test_pearson", torchmetrics.functional.pearson_corrcoef(logits.squeeze(), y.squeeze()))
-
-    def predict_step(self, batch, batch_idx):
-        x = batch
-        logits = self(x)
-
-        return logits.squeeze()
-
-    def configure_optimizers(self):
-        optimizer = torch.optim.AdamW(self.parameters(), lr=self.lr)
-        return optimizer
-
-
 if __name__ == '__main__':
-    # 하이퍼 파라미터 등 각종 설정값을 입력받습니다
-    # 터미널 실행 예시 : python3 run.py --batch_size=64 ...
-    # 실행 시 '--batch_size=64' 같은 인자를 입력하지 않으면 default 값이 기본으로 실행됩니다
     parser = argparse.ArgumentParser()
-    parser.add_argument('--model_name', default='klue/roberta-large', type=str)
+    parser.add_argument('--model_name', default='klue/roberta-small', type=str)
     parser.add_argument('--batch_size', default=16, type=int)
     parser.add_argument('--max_epoch', default=1, type=int)
     parser.add_argument('--shuffle', default=True)
@@ -225,30 +32,66 @@ if __name__ == '__main__':
     parser.add_argument('--test_path', default='../data/dev.csv')
     parser.add_argument('--predict_path', default='../data/test.csv')
     parser.add_argument('--num_workers', default=1, type=int)
+    parser.add_argument('--do_optimize', default=False)
+    parser.add_argument('--n_trials', default=1, type=int)
+
     args = parser.parse_args()
+    
+    wandb_logger = pl.WandbLogger(
+        project="wandb project",
+        name="wandb run name",
+        log_model="all",
+        save_dir="wandb"
+    )
 
-    # Optimize hyperparameters using optuna
     study = optuna.create_study(direction='minimize')
-    study.optimize(lambda trial: objective(trial, args), n_trials=50)
 
-    # Use best hyperparameters to train the final model
-    best_params = study.best_params
-    batch_size = best_params['batch_size'] or args.batch_size
-    max_epoch = best_params['max_epoch'] or args.max_epoch
+    dataloader = Dataloader(
+        args.model_name, 
+        args.batch_size,
+        args.shuffle,
+        args.train_path,
+        args.dev_path,
+        args.test_path,
+        args.predict_path,
+        args.num_workers
+    )
+    if args.do_optimize:
+        def objective(trial, args):
+            learning_rate = trial.suggest_float('learning_rate', 1e-6, 1e-4, log=True)
+            
+            model = Model(args.model_name, lr=learning_rate)
+            trainer = pl.Trainer(
+                accelerator="gpu", devices=1, 
+                max_epochs=args.max_epoch,
+                log_every_n_steps=1,
+                logger=False,
+                enable_checkpointing=False
+            )
+            
+            trainer.fit(model=model, datamodule=dataloader)
+            val_metrics = trainer.validate(model=model, datamodule=dataloader, verbose=False)
+            val_loss = val_metrics[0]['val_loss']
+            return val_loss
+
+        study.optimize(lambda trial: objective(trial, args), n_trials=args.n_trials)
+
+        best_params = study.best_params
     learning_rate = best_params['learning_rate'] or args.learning_rate
 
 
-    # dataloader와 model을 생성합니다.
-    dataloader = Dataloader(args.model_name, batch_size, args.shuffle, args.train_path, args.dev_path,
-                            args.test_path, args.predict_path, args.num_workers)
     model = Model(args.model_name, learning_rate)
 
-    # gpu가 없으면 accelerator="cpu"로 변경해주세요, gpu가 여러개면 'devices=4'처럼 사용하실 gpu의 개수를 입력해주세요
-    trainer = pl.Trainer(accelerator="gpu", devices=1, max_epochs=max_epoch, log_every_n_steps=1)
+    trainer = pl.Trainer(
+        accelerator="gpu", 
+        devices=1, 
+        max_epochs=args.max_epoch, 
+        log_every_n_steps=1, 
+        logger=wandb_logger
+    )
 
     # Train part
     trainer.fit(model=model, datamodule=dataloader)
     trainer.test(model=model, datamodule=dataloader)
 
-    # 학습이 완료된 모델을 저장합니다.
     torch.save(model, 'model.pt')
